Add tests (#16)
This PR adds tests of the logger and runs them as part of the conda
package build. To ensure correctness of both the code and the conda
package itself, the tests are built and run both as part of the package
build and in a subsequent test stage against the installed package. The
wheel is currently untested since there is not an easy way to do that
and we don't test C++ wheels anywhere else in RAPIDS currently. If we
start doing that elsewhere we can port the approach to here.