Enable logger macro generation  (#19)
This PR adds the ability to generate project-specific logger macros. It
also adds tests for this use case. Since these macros are generated via
CMake at project configure time, the new tests necessitated a new test
harness for testing CMake builds as part of the test suite. That is now
added as well.