feat(cmake) This commit adds packaging
The platform pipelines requires projects to produce artifacts,
so this commit adds a cpack option to package the logger into
either a debian, tgz, or rpm package.

We have added a cmake config file template so that the logger
can be found in projects with a `find_package(RAPIDS_LOGGER CONFIG REQUIRED)
`
in a CMakeLists.txt file.

During installation, we only need the function definition of
rapids_make_logger and cmake template files, not the rest
 of the root CMakeLists.txt file.
We have split up the root CMakeList.txt file into itself and
rapids_make_logger.cmake. So that we don't have to install
unnecessary cmake details for consumers.

Since amd-integration/1.0.x will not receive any updates,
we have decided to diverge from the upstream project to
make usage easier.